어드민 로그인 시 어드민 페이지 연결
어드민 계정으로 로그인시 어드민 페이지로 넘어갈 수 있도록 화면 연결

+++ momo_project/src/main/java/com/momo/admin/AdminMemberRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.momo.domain.Member;
 
-public interface AdminMemberRepository extends JpaRepository<Member, Long>{
+public interface AdminMemberRepository extends JpaRepository<Member, Integer>{
 
 }
